IDE-1095 Ports returning the UnknownPort exception shouldnt stop other ports from displaying
Previously, if the IStatus returned to to the MergePortsCommand was not perfect, all ports were unset
from the port container. The MergePortsCommand was changed to allow for any status with a severity
less than error.

The ScaAbstractComponentImpl which previously treated the UnknownPortException as an error will now
treat it as a warning.