Updates and improvements to the resource factory
We now use absolute file path for the profile name, both for resources in the
SDRROOT and the workspace. We maintain a special profile name for resources
supplied by bundles.

Because the file manager provides the local file system's contents by default,
there is no need to mount a file system for resources from the SDRROOT or
workspace. This greatly reduces the amount of indirection that occurs during
file system access.

Dropped the ability to specify a priority for different resource factories.
Resources with the same profile name would have shadowed each other, creating
confusion for users about which resource was actually running. It also added a
layer of CORBA indirection in getting to the file system. Now the resource
factory will report an error if mounting a resource would shadow an existing
mount.