Couldn't connect to two domains with the same name
This problem presented any time a user connected to two domains with the same
name, but using different naming services (likely on two different machines).
The IDE caches ORB settings, the ORB itself, and the root POA used for a given
domain. They were getting cached using the domain name as the key, which isn't
unique since we started allowing multiple domains with the same name.

Thus, connecting to domain A via naming service 1 caused that naming service
info to get cached, and when you tried to connect to domain A via naming
service 2, the information for naming service 2 didn't get used; the cached
info for naming service 1 got used instead.

Now the domain label is used as the key (the displayed name, which is unique).