Mitigate performance issues while browsing file system
There were several issues found. First, child file stores were being added to
their parent one at a time rather than as a group, which triggered an
individual notification for each.

Second, the item provider had poor performance if a large number of children
were revealed at the same time. Every file store's content was being downloaded
to determine what icon should be displayed next to it. First, the number of
simultaneous jobs was limited so the thread count doesn't spike. Next the
provider was modified to determine the icon based on filename alone for large
files. Lastly, the default icon is set based on filename, and no update is
performed if the icon is determined to be the same. This helps prevent
unnecessary viewer refreshes.