IDE-1824 Prevent mounting over existing file system locations
An IResourceFactoryProvider could supply a mount point which happened to match
an existing filed/directory in the file system. The mount point took
precedence, shadowing the existing file/directory. Now the
ResourceFactoryRegistry checks and only mounts the file system if the path
does not already exist, or if it does is an empty directory.