IDE-1911 Correct some SCA model problems when CORBA obj is null
ScaService had most of the issues. Many of these were due to the fact that the
model must dynamically decide whether a particular service implements the right
IDL for an operation, and then narrow the CORBA object. Those objects were
cached for speed purposes, and needed to be more in sync with the service's
main CORBA object.