IDE-1911 Don't try to configure remote CORBA object for design-time editing
The class providing editing for properties of resources in SAD/DCD files was
not setting the appropriate flag to suppress the code that tried to update the
remote object via CORBA call.

The code is a bit simpler now as well, and can suppress work due to multiple
calls to setInput() with the same object (there appear to be two calls in our
diagrams).