IDE-1955 Ensure transaction on Resource while reading
If a Resource change occurred in a non-UI thread, the code made an asyncExec()
to update the XML document's text. The code was then running outside the
noticiation, and hence outside the context of a transaction. This caused odd
synchronization issues between the EMF model and the XML editor's text.

This is an important fix for anywhere the EMF model being used by an editor
might be modified outside the UI thread.