IDE-942 Make FEI tuner alloc property writeonly
The intention is to avoid confusing the user. When the property was readwrite,
it was queryable, and thus showed in the properties view. But you didn't have
any defined behavior, and usually it just showed the last allocation request,
regardless of whether it was successful or if the actual fulfillment of the
request was slightly different than what was requested.