IDE-2038 Cleanup handler; correct issue with multiple SRI listeners on a tuner
Previously if someone with another IDE displayed the SRI for a tuner, the IDE
assumed that meant the current IDE was already viewing the SRI and would not
displayed it. The new algorithm isn't perfect but is better - it looks to see
if a view is already open showing SRI for that port.