IDE-2080 Ensure model properties can correctly apply property overrides
The previous method involved doing something like:

  ScaProp.fromAny(PropRef.toAny())

That transition essentially went like this:

  String -> Native Java Type -> Any -> Native Java Type

This is not only excessive, but doesn't operate correctly unless there is
symmetry in the mapping between String to Any. That is, for any String, we must
be able to convert to an Any, and then back to the same String without any
changes. With the introduction of using "now" to denote current time, the
mapping no longer holds. Once "now" is converted to the current time inside an
Any, we can't convert it back to the word "now", because that information has
been lost.

The transition is now:

  String -> Native Java Type

This will allow us to preserve "now" by using a native type that can store that
information.