IDE-2097 Fix clearing simple sequence with null, stop hiding errors
If an Any containing anything other than a sequence (array) was received, it
was being ignored. Now, it should cause an exception which will be marked as
a runtime error against the property value. Also, an Any representing a null
should cause the value of the simple sequence to be set to null.