Handle corner case when setting value of props in model based on query()
A corner case existed where the properties were queried before their PRF
definitions were available. In this case the IDE creates the SCA model props
based on the types returned from query(). Later, when the PRF definitions were
available, the IDE was trying to add them to the existing properties, but there
was a type mismatch (simple vs simple-sequence), which caused a
ClassCastException.