Minor corrections for the CommonNavigator in the explorer view...
In a few places we inject fake container objects between objects in the model
hierarchy. For example, under the domain (ScaDomainManager) we inject "Device
Managers", "Event Channels" and "Waveforms" folders. We represent these by
using the item provider for that object as the model object.

This injection must be declared to the CommonNavigator. An entry in the
plugin.xml provides this info. Otherwise, when tracing from a leaf node back
towards the root of the tree the navigator is unable to make the jump from the
injected item to its parent. This causes problems locating items within the
tree. This was a problem for the "External Ports" folder on waveforms.

Separately, the check on device managers to ensure their content is present
when they get expanded didn't include a check that services had been fetched.
While unlikely to occur in reality, it is nonetheless corrected.