IDE-2137 Correct xcore error
The correct syntax in Xcore to specify a zero-to-many constraint is like so:

  Object[]

Previously we were using the following:

  EEList<Object>

and importing:

  import org.eclipse.emf.ecore.EEList

That's a sort of backwards way to arrive at the desired effect. In any case the
two operations which were doing this aren't used (of course). So the code is
just being deleted.