IDE 2178 Allow scanner allocation via FEI runtime allocation wizard
This allows both design-time (i.e. FEI uses device in a SAD diagram) and
runtime allocation of a scanner. In the FEI allocation wizard, if the user
chooses RX_SCANNER_DIGITIZER, an extra page is added to the wizard that allows
the user to provide the values for the FRONTEND::scanner_allocation property.