IDE-2149 NPE when opening adv. plot wiz. from an SDDS port
This commit also corrects an omission from IDE-2043. The advanced plot wizard
for an SDDS port now also supports stream selection if the SDDS port is
multi-out.

The validation has been modified to only present a warning if the user clears
the connection ID. No connection ID indicates the UI should generate a
connection ID, and the UI shouldn't view this any differently than the user
typing in a connection ID that isn't in the list of choices (already a
warning).

The UI still shows an error (and thus prevents finishing the wizard) if the
user chooses a connection ID that is already in use.