IDE-1713 Prevent save error dialog
The dialog was popping because the previoud IDE-1713 code change caused the
SAD/DCD editors to start validating the gdiagram files. Directly editing the
SAD/DCD XML can apparently still cause some link issues in the diagram, and the
save was catching those. We shouldn't present those to the user, because it's
not something they (really) did, nor something we'd expect them to act on.