IDE-2201 Don't refresh children of dom mgr / dev mgr
We will no longer refresh the children of dom mgr / dev mgrs from the polling
data provider. This corrects two issues:
* The dom mgr / dev mgr refresh was long (*especially* dom mgr)
* The children ended up being refreshed twice (via parent and self)

The downside to this is this means we won't discover new children / removed
children (existing children stay up-to-date). However, we can let the
ODM_Channel data provider handle this job, which it's already doing much more
promptly and efficiently. The only thing it doesn't handle is discovering
event channels, so we retain that as a special case of the refresh logic for
the domain manager.