IDE-2204 Ensure CORBA object is released
Previously we watched the field corbaObj (the narrowed object), but that isn't
guaranteed to be set. The obj field gets set first, then corbaObj later after
the object is narrowed. By watching the obj field, we'll guarantee we release
the object when obj is changed/unset, even if fetchNarrowedObject() was never
called.

We also go straight to CorbaUtils.release(obj) rather than going through an
Eclipse Job, which didn't give us anything but a more delayed and indirect
release.