IDE-2292 Correct event data provider's add/remove
This commit corrects a few issues with the code that handles add/remove events
from the domain manager's ODM_Channel.

First - when handling add events the code now creates the SCA model objects
with all the information it has available to it from the event. This includes
the identifier and the label/name for most objects. This is important because
if the object has a very short lifespan the IDE might not have time to perform
a refresh of the model object and obtain the identifier and label/name. When a
remove event is received for the object, the IDE has to find the object based
on the identifier (only the identifier and label/name are available in the
event).

Second - when handling add events the IDE first checks if the container for the
object has been set or not. If it isn't set, this means the IDE has never
fetched data about those objects, and we do NOT add the object to the model.
For example, if the IDE gets an event about a new device manager, but the
ScaDomainManager model object's list of device managers isn't set (i.e.
!ScaDomainManager.isSetDeviceManagers()) then the IDE has never fetched that
list. If we were to add the new device manager described in the event, then the
list would be marked set and the IDE would think it didn't need to fetch it
when the user browsed it in the UI. This could potentially cause many device
managers to not be included in the model (or displayed).

Third - in a few places the code need to use our EMF model code for protecting
against concurrent modifications.