IDE-2292 Use refresh job when event doesn't have enough info
The event data provider now uses a job that performs a very selective refresh
when the information in an event can't be used to located the object to be
removed. The jobs are also used in the case of adding devices/services since
there's no way to know which device manager parents the device/service.